Extract more code from plugin driver
